smartbudget_ai/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
python-dotenv
mysql-connector-python
httpx
